game join money validation works

+++ db/migrate/20200423184042_create_users.rb
@@ -5,7 +5,7 @@ class CreateUsers < ActiveRecord::Migration[6.0]
       t.string :password_digest
       t.string :bio
       t.string :avatar
-      t.integer :winnings, :default => 0
+      t.float :winnings, :default => 0
       t.float :deposit, :default => 0
       t.integer :number_wins, :default => 0
